chore: Add crane pipeline that checks image environment
This would be used to compare values for variables we set across our
packages to ensure there is not drift between pinned dependencies in our
packaging and in upstream images

There are a lot of images that intentionally ship pinned dependencies. A
number of those images also declare the version of those dependencies
within the image's environment. This is commonly seen across AI/ML
images

Limited to assessing env vars for now, but we can consolidate everything
into a go module later for validating labels, ensuring entrypoints are
provided in packages at the correct location for packages that are
supposed to provide them, etc

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>