default binary name to package name in help and ver-check
In really a good amount of cases in wolfi, the binary output matches
the package name.
this commit changes the defaults of help-check and ver-check inputs.

with this change, in a good amount of packages, we can use the pipeline
in following manner:
```bash
- uses: test/tw/ver-check
- uses: test/tw/help-check
```

the change is backwards compatible because we've pinned things everywhere
till now so existing pipeline usage is not affected by this change.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>

Co-authored-by: Dentrax <furkan.turkal@chainguard.dev>