chore: remove allow-dangling from symlink-check pipelines
there are other ways to verify the symlink like stat or readlink.
the flag `--allow-absolute` is still there but I'm inclined towards
not exposing it at pipelines level.

If we use this at inside package config then the whole point of having
a symlink-check is broken. We want to validate the integrity of fs
by making sure there's no broken symlink.

By exposing this, we're entering a greyish area where we're allowing
this and bots are learning the same.
Not exposing seems like a better default than what we had previously.
Happy to discuss other thoughts here as well.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>